Fix TDT loss failing on int32 targets
torch.gather only accepts an int64 index before torch 2.8, and
ParakeetForTDTLoss casts the targets to int32 before handing them to
tdt_loss. Cast to int64 where the tensor becomes a gather index, so that
both the model's labels= path and a direct tdt_loss call with int32
targets work on every torch version the setup declares support for.

+++ src/transformers/loss/loss_tdt.py
@@ -80,7 +80,9 @@ def tdt_loss(
     blank_log_probs = token_log_probs[:, :, :, blank_token_id]
 
     if max_u > 1:
-        targets_expanded = targets.unsqueeze(1).expand(-1, max_t, -1)  # (batch, T, U_labels)
+        # torch.gather only accepts an int64 index before torch 2.8, and callers
+        # legitimately pass int32 targets (ParakeetForTDTLoss casts to int32 itself).
+        targets_expanded = targets.long().unsqueeze(1).expand(-1, max_t, -1)  # (batch, T, U_labels)
         label_log_probs = torch.gather(
             token_log_probs[:, :, : max_u - 1, :],  # (batch, T, U-1, vocab)
             dim=3,
